Update handling of errors.
The core part of this patch is to remove 'cause' attribute
from GaxError. Actually Ruby Exception already has such method
from 2.1.0, and $ERROR_INFO contains the previous exception
(or cause) for an exception.

As the consequence of this redesign, we will have to raise the
RetryError inside of rescue clause (otherwise $ERROR_INFO is
not set), and therefore retryable is now an infinite loop.

Also removed an unnecessary if-clause for catching errors,
instead bringing them into the matcher of rescue clause.

And fix some other lint errors:
- raise Error.new(msg) => raise Error, msg -- recommended
  style from rubocop.
- avoid using raise msg -- from internal feedback
- fix some typos

fixes #16