Hide CallSettings as private.
In Ruby, a 'private constant' is a constant which will have
to be accessed without prefix. So all entties inside of
Google::Gax can still access it as-is, but outside the module
it can't be accessed as Google::Gax::CallSettings.

This affects the tests, so this patch introduced a renaming
to toplevel ::CallSettings in a way to avoid violating
access of private constant.

updates https://github.com/googleapis/gax-python/issues/112