Make headers on operations client injectable
For Google Ads we send additional headers for LRO reloads, specifically:

```
developer-token
login-customer-id
```

without these headers being injectable it's not possible for us to
complete LRO calls successfully. This patch exposes headers in the
constructor of the operations client so that we can pass these when we
construct our operations client objects.